+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Versions
 
+## 1.7.3
+
+- Show L-BTC balance changes when sending backup by telegram
+- Add generating bech32m addresses for descriptors wallets
+
 ## 1.7.2
 
 - Ignore inline comments in peerswap.conf